==> analyses/mol-clinical-annotation-files/code/05_adding_ancestry.ipynb <==
import pandas as pd 
import numpy as np

# This file has all the samples that have ancestry information (only ones that have germline VCF files)
ancestry_file = pd.read_csv("ancestry_check/output/CBTTC_and_PNOC003.txt", sep="\t")

# removing row 35 with "ancestry" \t "case_id"
ancestry_file = ancestry_file[ancestry_file['ancestry'] != 'ancestry']

#Renaming column to cohort_participant_id so we can use merge with annotation files
ancestry_file = ancestry_file.rename(columns={"case_id": "cohort_participant_id"})

annot1 = pd.read_csv("output/annotation_files/pbta-histologies.tsv", sep="\t")

# This fucntio will take annotation and ancestry file with "cohort_participant_id" and merge those
# Sin ce we are doing an outer merge, any sample with no ancestry will have "NA" as value 
# If peddy was not able to determine population, value will be "UNKNOWN"
# It will also remove PNOC008 samples 
def merge_ancestry_and_annotation(ancestry_df, annotation_df):
    merged_df = pd.merge(annotation_df,ancestry_df, on=[
    'cohort_participant_id'], how="outer").fillna("NA")
    #merged_df = merged_df[merged_df["cohort"]  != "PNOC008"]
    return(merged_df)

# Create new annotation files with ancestry and no PNOC008 samples 
annot1_out = merge_ancestry_and_annotation( annot1,ancestry_file)


annot1_out.to_csv("output/annotation_files/pbta-histologies.tsv", sep="\t",index=False)
